requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Deep Learning and Computer Vision
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
opencv-python==4.9.0.80
numpy>=1.24.3,<2.0

# Firebase (Firestore Database)
firebase-admin>=6.2.0

# PromptLayer for RAG evaluation
promptlayer==0.3.5

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml>=5.1

# CORS
fastapi-cors==0.0.6

# Image processing
imageio==2.33.1
imageio-ffmpeg==0.4.9
